restructure sbk-gem and sbk-gem-yal packages.
