+++ checkstyle/import-control.xml
@@ -24,6 +24,7 @@
     <allow pkg="io.sbp"/>
     <allow pkg="io.sbk"/>
     <allow pkg="io.sbm"/>
+    <allow pkg="io.gem"/>
     <allow pkg="io.pravega"/>
     <allow pkg="lombok"/>
     <allow pkg="ch.qos.logback"/>
