--- sbk-gem/src/main/java/io/sbk/logger/GemLogger.java
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) KMG. All Rights Reserved..
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- */
-
-package io.sbk.logger;
-
-import io.sbk.params.GetArguments;
-import io.sbm.logger.RamLogger;
-
-/**
- * Interface GemLogger.
- */
-public interface GemLogger extends RamLogger, GetArguments {
-}
