--- sbk-gem/src/main/java/io/sbk/logger/impl/AbstractGemLogger.java
@@ -1,22 +1,0 @@
-/**
- * Copyright (c) KMG. All Rights Reserved..
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- */
-
-package io.sbk.logger.impl;
-
-import io.sbk.logger.GemLogger;
-import io.sbm.logger.impl.SbmPrometheusLogger;
-
-public abstract class AbstractGemLogger extends SbmPrometheusLogger implements GemLogger {
-    @Override
-    public abstract String[] getOptionsArgs();
-
-    @Override
-    public abstract String[] getParsedArgs();
-}
