--- sbk-gem/src/main/java/io/sbk/params/GemParameterOptions.java
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) KMG. All Rights Reserved..
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- */
-
-package io.sbk.params;
-
-/**
- * Interface GemParameterOptions.
- */
-public non-sealed interface GemParameterOptions extends InputParameterOptions, GemParameters, GetArguments {
-
-}
-
